Shuffle header files
This has significant effects on the compile time of webfort.cpp, but
server.cpp still has the bad end of the stick.
