+++ server/input.hpp
@@ -1,0 +1,85 @@
+/* FIXME: input handling is long-winded enough to get its own file. */
+#include "SDL_events.h"
+#include "SDL_keysym.h"
+extern "C"
+{
+extern int SDL_PushEvent( SDL::Event* event );
+}
+static SDL::Key mapInputCodeToSDL( const uint32_t code )
+{
+#define MAP(a, b) case a: return b;
+    switch (code)
+    {
+    // {{{ keysyms
+    MAP(96, SDL::K_KP0);
+    MAP(97, SDL::K_KP1);
+    MAP(98, SDL::K_KP2);
+    MAP(99, SDL::K_KP3);
+    MAP(100, SDL::K_KP4);
+    MAP(101, SDL::K_KP5);
+    MAP(102, SDL::K_KP6);
+    MAP(103, SDL::K_KP7);
+    MAP(104, SDL::K_KP8);
+    MAP(105, SDL::K_KP9);
+    MAP(144, SDL::K_NUMLOCK);
+
+    MAP(111, SDL::K_KP_DIVIDE);
+    MAP(106, SDL::K_KP_MULTIPLY);
+    MAP(109, SDL::K_KP_MINUS);
+    MAP(107, SDL::K_KP_PLUS);
+
+    MAP(33, SDL::K_PAGEUP);
+    MAP(34, SDL::K_PAGEDOWN);
+    MAP(35, SDL::K_END);
+    MAP(36, SDL::K_HOME);
+    MAP(46, SDL::K_DELETE);
+
+    MAP(112, SDL::K_F1);
+    MAP(113, SDL::K_F2);
+    MAP(114, SDL::K_F3);
+    MAP(115, SDL::K_F4);
+    MAP(116, SDL::K_F5);
+    MAP(117, SDL::K_F6);
+    MAP(118, SDL::K_F7);
+    MAP(119, SDL::K_F8);
+    MAP(120, SDL::K_F9);
+    MAP(121, SDL::K_F10);
+    MAP(122, SDL::K_F11);
+    MAP(123, SDL::K_F12);
+
+    MAP(37, SDL::K_LEFT);
+    MAP(39, SDL::K_RIGHT);
+    MAP(38, SDL::K_UP);
+    MAP(40, SDL::K_DOWN);
+
+    MAP(188, SDL::K_LESS);
+    MAP(190, SDL::K_GREATER);
+
+    MAP(13, SDL::K_RETURN);
+
+    //MAP(16, SDL::K_LSHIFT);
+    //MAP(17, SDL::K_LCTRL);
+    //MAP(18, SDL::K_LALT);
+
+    MAP(27, SDL::K_ESCAPE);
+#undef MAP
+    // }}}
+    }
+    if (code <= 177)
+        return (SDL::Key)code;
+    return SDL::K_UNKNOWN;
+}
+
+void simkey(int down, int mod, SDL::Key sym, int unicode)
+{
+    SDL::Event event;
+    memset(&event, 0, sizeof(event));
+
+    event.type = down ? SDL::ET_KEYDOWN : SDL::ET_KEYUP;
+    event.key.state = down ? SDL::BTN_PRESSED : SDL::BTN_RELEASED;
+    event.key.ksym.mod = (SDL::Mod)mod;
+    event.key.ksym.sym = sym;
+    event.key.ksym.unicode = unicode;
+
+    SDL_PushEvent(&event);
+}
